requirements.txt
# Production dependencies
streamlit>=1.31.0
boto3>=1.34.34
Pillow>=10.2.0
pydantic>=2.6.1
python-dotenv>=1.0.1
requests>=2.31.0

# AI/ML dependencies
transformers>=4.36.0
torch>=2.1.0
torchvision>=0.16.0

# AWS SDK extras
boto3-stubs[essential]==1.34.34

# Testing dependencies
pytest==8.0.0
pytest-cov==4.1.0
pytest-asyncio==0.23.4
moto[all]==5.0.0

# Code quality
black==24.1.1
flake8==7.0.0
mypy==1.8.0

# CI/CD and utilities
awscli==1.32.34
python-jose[cryptography]==3.3.0